+++ nmosnode/aggregator.py
@@ -17,8 +17,9 @@ from gevent import monkey
 monkey.patch_all()
 
 from six import itervalues # noqa E402
-from six.moves.urllib.parse import urljoin # noqa E402
+from six.moves.urllib.parse import urljoin, urlparse # noqa E402
 
+from collections import deque # noqa E402
 import requests # noqa E402
 import json # noqa E402
 import time # noqa E402
@@ -26,40 +27,33 @@ import time # noqa E402
 import gevent.queue # noqa E402
 
 from nmoscommon.logger import Logger # noqa E402
-from mdnsbridge.mdnsbridgeclient import IppmDNSBridge # noqa E402
+from mdnsbridge.mdnsbridgeclient import IppmDNSBridge, NoService, EndOfServiceList # noqa E402
 from nmoscommon.mdns.mdnsExceptions import ServiceNotFoundException # noqa E402
 
 from nmoscommon.nmoscommonconfig import config as _config # noqa E402
 import traceback # noqa E402
 
-AGGREGATOR_APIVERSION = _config.get('nodefacade').get('NODE_REGVERSION')
 AGGREGATOR_APINAMESPACE = "x-nmos"
 AGGREGATOR_APINAME = "registration"
 
 LEGACY_REG_MDNSTYPE = "nmos-registration"
 REGISTRATION_MDNSTYPE = "nmos-register"
 
-
-class NoAggregator(Exception):
-    def __init__(self, mdns_updater=None):
-        if mdns_updater is not None:
-            mdns_updater.inc_P2P_enable_count()
-        super(NoAggregator, self).__init__("No Registration API found")
+BACKOFF_INITIAL_TIMOUT_SECONDS = 5
+BACKOFF_MAX_TIMEOUT_SECONDS = 40
 
 
 class InvalidRequest(Exception):
-    def __init__(self, status_code=400, mdns_updater=None):
-        if mdns_updater is not None:
-            mdns_updater.inc_P2P_enable_count()
+    """Client Side Error during request, HTTP 4xx"""
+    def __init__(self, status_code=400):
         super(InvalidRequest, self).__init__("Invalid Request, code {}".format(status_code))
         self.status_code = status_code
 
 
-class TooManyRetries(Exception):
-    def __init__(self, mdns_updater=None):
-        if mdns_updater is not None:
-            mdns_updater.inc_P2P_enable_count()
-        super(TooManyRetries, self).__init__("Too many retries.")
+class ServerSideError(Exception):
+    """Exception raised when a HTTP 5xx, timeout or inability to connect returned during request.
+    This indicates a server side or connectivity issue"""
+    pass
 
 
 class Aggregator(object):
@@ -69,12 +63,14 @@ class Aggregator(object):
     def __init__(self, logger=None, mdns_updater=None):
         self.logger = Logger("aggregator_proxy", logger)
         self.mdnsbridge = IppmDNSBridge(logger=self.logger)
-        self.aggregator = ""
+        self.aggregator_apiversion = None
+        self.service_type = None
+        self._set_api_version_and_srv_type(_config.get('nodefacade').get('NODE_REGVERSION'))
+        self.aggregator = None
         self.registration_order = ["device", "source", "flow", "sender", "receiver"]
         self._mdns_updater = mdns_updater
-        # 'registered' is a local mirror of aggregated items. There are helper methods
-        # for manipulating this below.
-        self._registered = {
+        # '_node_data' is a local mirror of aggregated items.
+        self._node_data = {
             'node': None,
             'registered': False,
             'entities': {
@@ -83,173 +79,138 @@ class Aggregator(object):
             }
         }
         self._running = True
+        self._aggregator_list_stale = True
+        self._aggregator_failure = False  # Variable to flag when aggregator has returned and unexpected error
+        self._backoff_active = False
+        self._backoff_period = 0
         self._reg_queue = gevent.queue.Queue()
-        self.heartbeat_thread = gevent.spawn(self._heartbeat)
+        self.main_thread = gevent.spawn(self._main_thread)
         self.queue_thread = gevent.spawn(self._process_queue)
 
-    # The heartbeat thread runs in the background every five seconds.
-    # If when it runs the Node is believed to be registered it will perform a heartbeat
-    def _heartbeat(self):
-        self.logger.writeDebug("Starting heartbeat thread")
-        while self._running:
-            heartbeat_wait = 5
-            if not self._registered["registered"]:
-                self._process_reregister()
-            elif self._registered["node"]:
-                # Do heartbeat
-                try:
-                    self.logger.writeDebug("Sending heartbeat for Node {}"
-                                           .format(self._registered["node"]["data"]["id"]))
-                    self._SEND("POST", "/health/nodes/" + self._registered["node"]["data"]["id"])
-                except InvalidRequest as e:
-                    if e.status_code == 404:
-                        # Re-register
-                        self.logger.writeWarning("404 error on heartbeat. Marking Node for re-registration")
-                        self._registered["registered"] = False
+    def _set_api_version_and_srv_type(self, api_ver):
+        """Set the aggregator api version equal to parameter and DNS-SD service type based on api version"""
+        self.aggregator_apiversion = api_ver
+        self._set_service_type(api_ver)
 
-                        if(self._mdns_updater is not None):
-                            self._mdns_updater.inc_P2P_enable_count()
-                    else:
-                        # Client side error. Report this upwards via exception, but don't resend
-                        self.logger.writeError("Unrecoverable error code {} received from Registration API on heartbeat"
-                                               .format(e.status_code))
-                        self._running = False
-                except Exception:
-                    # Re-register
-                    self.logger.writeWarning("Unexpected error on heartbeat. Marking Node for re-registration")
-                    self._registered["registered"] = False
+    def _set_service_type(self, api_ver):
+        """Set DNS-SD service type based on current api version in use"""
+        if api_ver in ['v1.0', 'v1.1', 'v1.2']:
+            self.service_type = LEGACY_REG_MDNSTYPE
+        else:
+            self.service_type = REGISTRATION_MDNSTYPE
+
+    def _main_thread(self):
+        """The main thread runs in the background.
+        If when it runs the Node is believed to be registered it will perform a heartbeat every 5 seconds.
+        If the Node is not registered it will try register the Node"""
+        self.logger.writeDebug("Starting main thread")
+
+        while self._running:
+            if self._node_data["node"] and self.aggregator is None:
+                self._discovery_operation()
+            elif self._node_data["node"] and self._node_data["registered"]:
+                self._registered_operation()
             else:
-                self._registered["registered"] = False
-                if(self._mdns_updater is not None):
-                    self._mdns_updater.inc_P2P_enable_count()
-            while heartbeat_wait > 0 and self._running:
-                gevent.sleep(1)
-                heartbeat_wait -= 1
+                self._node_data["registered"] = False
+                self.aggregator = None
+                gevent.sleep(0.2)
+
         self.logger.writeDebug("Stopping heartbeat thread")
 
-    # Provided the Node is believed to be correctly registered, hand off a single request to the SEND method
-    # On client error, clear the resource from the local mirror
-    # On other error, mark Node as unregistered and trigger re-registration
-    def _process_queue(self):
-        self.logger.writeDebug("Starting HTTP queue processing thread")
-        # Checks queue not empty before quitting to make sure unregister node gets done
-        while self._running or (self._registered["registered"] and not self._reg_queue.empty()):
-            if not self._registered["registered"] or self._reg_queue.empty():
-                gevent.sleep(1)
-            else:
-                try:
-                    queue_item = self._reg_queue.get()
-                    namespace = queue_item["namespace"]
-                    res_type = queue_item["res_type"]
-                    res_key = queue_item["key"]
-                    if queue_item["method"] == "POST":
-                        if res_type == "node":
-                            data = self._registered["node"]
-                            try:
-                                self.logger.writeInfo("Attempting registration for Node {}"
-                                                      .format(self._registered["node"]["data"]["id"]))
-                                self._SEND("POST", "/{}".format(namespace), data)
-                                self._SEND("POST", "/health/nodes/" + self._registered["node"]["data"]["id"])
-                                self._registered["registered"] = True
-                                if self._mdns_updater is not None:
-                                    self._mdns_updater.P2P_disable()
+    def _discovery_operation(self):
+        """In Discovery operation the Node will wait a backoff period if defined to allow aggregators to recover when in
+        a state of error. Selecting the most appropriate aggregator and try to register with it.
+        If a registration fails then another aggregator will be tried."""
+        self.logger.writeDebug("Entering Discovery Mode")
 
-                            except Exception:
-                                self.logger.writeWarning("Error registering Node: %r" % (traceback.format_exc(),))
+        # Wait backoff period
+        # Do not wait backoff period if aggregator failed, a new aggregator should be tried immediately
+        if not self._aggregator_failure:
+            self._back_off_timer()
 
-                        elif res_key in self._registered["entities"][namespace][res_type]:
-                            data = self._registered["entities"][namespace][res_type][res_key]
-                            try:
-                                self._SEND("POST", "/{}".format(namespace), data)
-                            except InvalidRequest as e:
-                                self.logger.writeWarning("Error registering {} {}: {}".format(res_type, res_key, e))
-                                self.logger.writeWarning("Request data: {}".format(data))
-                                del self._registered["entities"][namespace][res_type][res_key]
+        self._aggregator_failure = False
 
-                    elif queue_item["method"] == "DELETE":
-                        translated_type = res_type + 's'
-                        try:
-                            self._SEND("DELETE", "/{}/{}/{}".format(namespace, translated_type, res_key))
-                        except InvalidRequest as e:
-                            self.logger.writeWarning("Error deleting resource {} {}: {}"
-                                                     .format(translated_type, res_key, e))
-                    else:
-                        self.logger.writeWarning("Method {} not supported for Registration API interactions"
-                                                 .format(queue_item["method"]))
-                except Exception:
-                    self._registered["registered"] = False
-                    if(self._mdns_updater is not None):
-                        self._mdns_updater.P2P_disable()
-        self.logger.writeDebug("Stopping HTTP queue processing thread")
+        # Update cached list of aggregators
+        if self._aggregator_list_stale:
+            self._flush_cached_aggregators()
 
-    # Queue a request to be processed.
-    # Handles all requests except initial Node POST which is done in _process_reregister
-    def _queue_request(self, method, namespace, res_type, key):
-        self._reg_queue.put({"method": method, "namespace": namespace, "res_type": res_type, "key": key})
+        while True:
+            self.aggregator = self._get_aggregator()
+            if self.aggregator is None:
+                self.logger.writeDebug("Failed to find registration API")
+                break
+            self.logger.writeDebug("Aggregator set to: {}".format(self.aggregator))
 
-    # Register 'resource' type data including the Node
-    # NB: Node registration is managed by heartbeat thread so may take up to 5 seconds!
-    def register(self, res_type, key, **kwargs):
-        self.register_into("resource", res_type, key, **kwargs)
+            # Perform initial heartbeat, which will attempt to register Node if not already registered
+            if self._heartbeat():
+                # Successfully registered Node with aggregator andproceed to registered operation
+                # Else will try next aggregator
+                break
 
-    # Unregister 'resource' type data including the Node
-    def unregister(self, res_type, key):
-        self.unregister_from("resource", res_type, key)
+    def _registered_operation(self):
+        """In Registered operation, the Node is registered so a heartbeat will be performed,
+        if the heartbeat is successful the Node will wait 5 seconds before attempting another heartbeat.
+        Else another aggregator will be selected"""
+        if not self._heartbeat():
+            # Heartbeat failed
+            # Flag to update cached list of aggregators and immediately try new aggregator
+            self.aggregator = None
+            self._aggregator_failure = True
 
-    # General register method for 'resource' types
-    def register_into(self, namespace, res_type, key, **kwargs):
-        data = kwargs
-        send_obj = {"type": res_type, "data": data}
-        if 'id' not in send_obj["data"]:
-            self.logger.writeWarning("No 'id' present in data, using key='{}': {}".format(key, data))
-            send_obj["data"]["id"] = key
-
-        if namespace == "resource" and res_type == "node":
-            # Handle special Node type
-            self._registered["node"] = send_obj
-        else:
-            self._add_mirror_keys(namespace, res_type)
-            self._registered["entities"][namespace][res_type][key] = send_obj
-        self._queue_request("POST", namespace, res_type, key)
-
-    # General unregister method for 'resource' types
-    def unregister_from(self, namespace, res_type, key):
-        if namespace == "resource" and res_type == "node":
-            # Handle special Node type
-            self._registered["node"] = None
-        elif res_type in self._registered["entities"][namespace]:
-            self._add_mirror_keys(namespace, res_type)
-            if key in self._registered["entities"][namespace][res_type]:
-                del self._registered["entities"][namespace][res_type][key]
-        self._queue_request("DELETE", namespace, res_type, key)
-
-    # Deal with missing keys in local mirror
-    def _add_mirror_keys(self, namespace, res_type):
-        if namespace not in self._registered["entities"]:
-            self._registered["entities"][namespace] = {}
-        if res_type not in self._registered["entities"][namespace]:
-            self._registered["entities"][namespace][res_type] = {}
-
-    # Re-register just the Node, and queue requests in order for other resources
-    def _process_reregister(self):
-        if self._registered.get("node", None) is None:
-            self.logger.writeDebug("No node registered, re-register returning")
-            return
-
+    def _heartbeat(self):
+        """Performs a heartbeat to registered aggregator
+        If heartbeat fails it will take actions to correct the error, by re-registering the Node
+        If successfull will return True, else will return False"""
+        if not self.aggregator:
+            return False
         try:
-            self.logger.writeDebug("Clearing old Node from API prior to re-registration")
-            self._SEND("DELETE", "/resource/nodes/" + self._registered["node"]["data"]["id"])
-        except InvalidRequest as e:
-            # 404 etc is ok
-            self.logger.writeInfo("Invalid request when deleting Node prior to registration: {}".format(e))
-        except Exception as ex:
-            # Server error is bad, no point continuing
-            self.logger.writeError("Aborting Node re-register! {}".format(ex))
-            return
+            R = self._send("POST", self.aggregator, self.aggregator_apiversion,
+                           "/health/nodes/{}".format(self._node_data["node"]["data"]["id"]))
 
-        self._registered["registered"] = False
-        if(self._mdns_updater is not None):
-            self._mdns_updater.inc_P2P_enable_count()
+            if R.status_code == 200 and self._node_data["registered"]:
+                # Continue to registered operation
+                self.logger.writeDebug("Successful heartbeat for Node {}"
+                                       .format(self._node_data["node"]["data"]["id"]))
+                self._registered()
+                heartbeat_wait = 5
+                while heartbeat_wait > 0 and self._running:
+                    gevent.sleep(1)
+                    heartbeat_wait -= 1
+                return True
+
+            elif R.status_code in [200, 409]:
+                # Delete node from registry
+                if self._unregister_node(R.headers.get('Location')):
+                    return self._register_node(self._node_data["node"])
+                else:
+                    # Try next registry
+                    return False
+
+        except InvalidRequest as e:
+            if e.status_code == 404:
+                # Re-register
+                self.logger.writeWarning("404 error on heartbeat. Marking Node for re-registration")
+                self._node_data["registered"] = False
+                return self._register_node(self._node_data["node"])
+            else:
+                # Other error, try next registry
+                return False
+        except ServerSideError:
+            self.logger.writeWarning("Server Side Error on heartbeat. Trying another registry")
+            return False
+        except Exception as e:
+            # Re-register
+            self.logger.writeWarning("Unexpected error on heartbeat. Marking Node for re-registration\n{}".format(e))
+            self._node_data["registered"] = False
+            return False
+
+    def _register_node(self, node_obj):
+        """Attempt to register Node with aggregator
+        Returns True is node was successfully registered with aggregator
+        Returns False if registration failed
+        If registration failed with 200 or 409, will attempt to delete and re-register"""
+        if node_obj is None:
+            return False
 
         # Drain the queue
         while not self._reg_queue.empty():
@@ -259,25 +220,40 @@ class Aggregator(object):
                 break
 
         try:
-            # Register the node, and immediately heartbeat if successful to avoid race with garbage collect.
-            self.logger.writeInfo("Attempting re-registration for Node {}"
-                                  .format(self._registered["node"]["data"]["id"]))
-            self._SEND("POST", "/resource", self._registered["node"])
-            self._SEND("POST", "/health/nodes/" + self._registered["node"]["data"]["id"])
-            self._registered["registered"] = True
-            if self._mdns_updater is not None:
-                self._mdns_updater.P2P_disable()
-        except Exception as e:
-            self.logger.writeWarning("Error re-registering Node: {}".format(e))
-            self.aggregator = ""  # Fallback to prevent us getting stuck if the Reg API issues a 4XX error incorrectly
-            return
+            # Try register the Node 3 times with aggregator before failing back to next aggregator
+            for i in range(0, 3):
+                R = self._send("POST", self.aggregator, self.aggregator_apiversion,
+                               "/resource", node_obj)
 
+                if R.status_code == 201:
+                    # Continue to registered operation
+                    self.logger.writeInfo("Node Registered with {} at version {}"
+                                          .format(self.aggregator, self.aggregator_apiversion))
+                    self._registered()
+
+                    # Trigger registration of Nodes resources
+                    self._register_node_resources()
+
+                    return True
+
+                elif R.status_code in [200, 409]:
+                    # Delete node from aggregator & re-register
+                    if self._unregister_node(R.headers.get('Location')):
+                        continue
+                    else:
+                        # Try next aggregator
+                        return False
+        except Exception as e:
+            self.logger.writeError("Failed to register node {}".format(e))
+        return False
+
+    def _register_node_resources(self):
         # Re-register items that must be ordered
         # Re-register things we have in the local cache.
         # "namespace" is e.g. "resource"
         # "entities" are the things associated under that namespace.
         for res_type in self.registration_order:
-            for namespace, entities in self._registered["entities"].items():
+            for namespace, entities in self._node_data["entities"].items():
                 if res_type in entities:
                     self.logger.writeInfo("Ordered re-registration for type: '{}' in namespace '{}'"
                                           .format(res_type, namespace))
@@ -288,7 +264,7 @@ class Aggregator(object):
         # Re-register things we have in the local cache.
         # "namespace" is e.g. "resource"
         # "entities" are the things associated under that namespace.
-        for namespace, entities in self._registered["entities"].items():
+        for namespace, entities in self._node_data["entities"].items():
             for res_type in entities:
                 if res_type not in self.registration_order:
                     self.logger.writeInfo("Unordered re-registration for type: '{}' in namespace '{}'"
@@ -296,91 +272,299 @@ class Aggregator(object):
                     for key in entities[res_type]:
                         self._queue_request("POST", namespace, res_type, key)
 
-    # Stop the Aggregator object running
-    def stop(self):
-        self.logger.writeDebug("Stopping aggregator proxy")
-        self._running = False
-        self.heartbeat_thread.join()
-        self.queue_thread.join()
+    def _registered(self):
+        """Mark Node as registered and reset counters"""
+        if(self._mdns_updater is not None):
+            self._mdns_updater.P2P_disable()
 
-    def status(self):
-        return {"api_href": self.aggregator,
-                "api_version": AGGREGATOR_APIVERSION,
-                "registered": self._registered["registered"]}
+        self._node_data['registered'] = True
+        self._aggregator_list_stale = True
 
-    def _get_api_href(self):
+        self._reset_backoff_period()
+
+    def _reset_backoff_period(self):
+        self.logger.writeDebug("Resetting backoff period")
+        self._backoff_period = 0
+
+    def _increase_backoff_period(self):
+        """Exponentially increase the backoff period, until set maximum reached"""
+        self.logger.writeDebug("Increasing backoff period")
+        self._aggregator_list_stale = True
+
+        if self._backoff_period == 0:
+            self._backoff_period = BACKOFF_INITIAL_TIMOUT_SECONDS
+            return
+
+        self._backoff_period *= 2
+        if self._backoff_period > BACKOFF_MAX_TIMEOUT_SECONDS:
+            self._backoff_period = BACKOFF_MAX_TIMEOUT_SECONDS
+
+    def _back_off_timer(self):
+        """Sleep for defined backoff period"""
+        self.logger.writeDebug("Backoff timer enabled for {} seconds".format(self._backoff_period))
+        self._backoff_active = True
+        gevent.sleep(self._backoff_period)
+        self._backoff_active = False
+
+    def _flush_cached_aggregators(self):
+        """Flush the list of cached aggregators in the mdns bridge client,
+        preventing the use of out of date aggregators"""
+        self.logger.writeDebug("Flushing cached list of aggregators")
+        self._aggregator_list_stale = False
+        self.mdnsbridge.updateServices(self.service_type)
+
+    def _get_aggregator(self):
+        """Get the most appropriate aggregator from the mdns bridge client.
+        If no aggregator found increment P2P counter, update cache and increase backoff
+        If reached the end of available aggregators update cache and increase backoff"""
+
         protocol = "http"
         if _config.get('https_mode') == "enabled":
             protocol = "https"
-        api_href = self.mdnsbridge.getHref(REGISTRATION_MDNSTYPE, None, AGGREGATOR_APIVERSION, protocol)
-        if api_href == "":
-            api_href = self.mdnsbridge.getHref(LEGACY_REG_MDNSTYPE, None, AGGREGATOR_APIVERSION, protocol)
-        return api_href
 
-    # Handle sending all requests to the Registration API, and searching for a new 'aggregator' if one fails
-    def _SEND(self, method, url, data=None):
-        if self.aggregator == "":
-            self.aggregator = self._get_api_href()
+        try:
+            return self.mdnsbridge.getHrefWithException(self.service_type, None, self.aggregator_apiversion, protocol)
+        except NoService:
+            self.logger.writeDebug("No Registration services found: {} {} {}".format(
+                                    self.service_type, self.aggregator_apiversion, protocol))
+            if self._mdns_updater is not None:
+                self._mdns_updater.inc_P2P_enable_count()
+            self._increase_backoff_period()
+            return None
+        except EndOfServiceList:
+            self.logger.writeDebug("End of Registration services list: {} {} {}".format(
+                                    self.service_type, self.aggregator_apiversion, protocol))
+            self._increase_backoff_period()
+            return None
 
-        headers = None
-        if data is not None:
-            data = json.dumps(data)
-            headers = {"Content-Type": "application/json"}
+    def _unregister_node(self, url_path=None):
+        """Delete node from registry, using url_path if specified"""
+        if self.aggregator is None:
+            self.logger.writeWarning('Could not un-register as no aggregator set')
+            return False
+        try:
+            self._node_data['registered'] = False
+            if url_path is None:
+                R = self._send('DELETE', self.aggregator, self.aggregator_apiversion,
+                               '/resource/nodes/{}'.format(self._node_data['node']["data"]["id"]))
+            else:
+                parsed_url = urlparse(url_path)
+                R = self._send_request('DELETE', self.aggregator, parsed_url.path)
 
-        url = AGGREGATOR_APINAMESPACE + "/" + AGGREGATOR_APINAME + "/" + AGGREGATOR_APIVERSION + url
-        for i in range(0, 3):
-            if self.aggregator == "":
-                self.logger.writeWarning("No aggregator available on the network or mdnsbridge unavailable")
-                raise NoAggregator(self._mdns_updater)
+            if R.status_code == 204:
+                # Successfully deleted node from Registry
+                self.logger.writeInfo("Node unregistered from {} at version {}"
+                                      .format(self.aggregator, self.aggregator_apiversion))
+                return True
+            else:
+                return False
+        except Exception as e:
+            self.logger.writeDebug('Exception raised while un-registering {}'.format(e))
+            return False
 
-            self.logger.writeDebug("{} {}".format(method, urljoin(self.aggregator, url)))
+    def _process_queue(self):
+        """Provided the Node is believed to be correctly registered, hand off a single request to the SEND method
+           On client error, clear the resource from the local mirror
+           On other error, mark Node as unregistered and trigger re-registration"""
+        self.logger.writeDebug("Starting HTTP queue processing thread")
+        # Checks queue not empty before quitting to make sure unregister node gets done
+        while self._running:
+            if (not self._node_data["registered"] or
+                    self._reg_queue.empty() or
+                    self._backoff_active or
+                    not self.aggregator):
+                gevent.sleep(1)
+            else:
+                try:
+                    queue_item = self._reg_queue.get()
+                    namespace = queue_item["namespace"]
+                    res_type = queue_item["res_type"]
+                    res_key = queue_item["key"]
+                    if queue_item["method"] == "POST":
+                        if res_type == "node":
+                            send_obj = self._node_data.get("node")
+                        else:
+                            send_obj = self._node_data["entities"][namespace][res_type].get(res_key)
 
-            # We give a long(ish) timeout below, as the async request may succeed after the timeout period
-            # has expired, causing the node to be registered twice (potentially at different aggregators).
-            # Whilst this isn't a problem in practice, it may cause excessive churn in websocket traffic
-            # to web clients - so, sacrifice a little timeliness for things working as designed the
-            # majority of the time...
-            try:
-                if _config.get('prefer_ipv6') is False:
-                    R = requests.request(method, urljoin(self.aggregator, url), data=data, timeout=1.0, headers=headers)
-                else:
-                    R = requests.request(method, urljoin(self.aggregator, url), data=data, timeout=1.0,
-                                         headers=headers, proxies={'http': ''})
-                if R is None:
-                    # Try another aggregator
-                    self.logger.writeWarning("No response from aggregator {}".format(self.aggregator))
+                        if send_obj is None:
+                            self.logger.writeError("No data to send for resource {}".format(res_type))
+                            continue
+                        try:
+                            self._send("POST", self.aggregator, self.aggregator_apiversion,
+                                       "/{}".format(namespace), send_obj)
+                            self.logger.writeInfo("Registered {} {} {}".format(namespace, res_type, res_key))
+                        except InvalidRequest as e:
+                            self.logger.writeWarning("Error registering {} {}: {}".format(res_type, res_key, e))
+                            self.logger.writeWarning("Request data: {}".format(send_obj))
+                            del self._node_data["entities"][namespace][res_type][res_key]
 
-                elif R.status_code in [200, 201]:
-                    if R.headers.get("content-type", "text/plain").startswith("application/json"):
-                        return R.json()
+                    elif queue_item["method"] == "DELETE":
+                        translated_type = res_type + 's'
+                        if namespace == "resource" and res_type == "node":
+                            # Handle special Node type
+                            self._node_data["node"] = None
+                            self._node_data["registered"] = False
+                        try:
+                            self._send("DELETE", self.aggregator, self.aggregator_apiversion,
+                                       "/{}/{}/{}".format(namespace, translated_type, res_key))
+                            self.logger.writeInfo("Un-registered {} {} {}".format(namespace, translated_type, res_key))
+                        except InvalidRequest as e:
+                            self.logger.writeWarning("Error deleting resource {} {}: {}"
+                                                     .format(translated_type, res_key, e))
                     else:
-                        return R.content
+                        self.logger.writeWarning("Method {} not supported for Registration API interactions"
+                                                 .format(queue_item["method"]))
+                except ServerSideError:
+                    self.aggregator = None
+                    self._aggregator_failure = True
+                    self._add_request_to_front_of_queue(queue_item)
+                except Exception as e:
+                    self.logger.writeError("Unexpected Error while processing queue, marking Node for re-registration\n"
+                                           "{}".format(e))
+                    self._node_data["registered"] = False
+                    self.aggregator = None
+                    if(self._mdns_updater is not None):
+                        self._mdns_updater.P2P_disable()
+        self.logger.writeDebug("Stopping HTTP queue processing thread")
 
-                elif R.status_code == 204:
-                    return
+    def _queue_request(self, method, namespace, res_type, key):
+        """Queue a request to be processed.
+           Handles all requests except initial Node POST which is done in _process_reregister"""
+        self._reg_queue.put({"method": method, "namespace": namespace, "res_type": res_type, "key": key})
 
-                elif (R.status_code//100) == 4:
-                    self.logger.writeWarning("{} response from aggregator: {} {}"
-                                             .format(R.status_code, method, urljoin(self.aggregator, url)))
-                    raise InvalidRequest(R.status_code, self._mdns_updater)
+    def _add_request_to_front_of_queue(self, request):
+        """Adds item to the front of the queue"""
 
-                else:
-                    self.logger.writeWarning("Unexpected status from aggregator {}: {}, {}"
-                                             .format(self.aggregator, R.status_code, R.content))
+        new_queue = deque()
+        new_queue.append(request)
 
-            except requests.exceptions.RequestException as ex:
-                # Log a warning, then let another aggregator be chosen
-                self.logger.writeWarning("{} from aggregator {}".format(ex, self.aggregator))
+        # Drain the queue
+        while not self._reg_queue.empty():
+            try:
+                new_queue.append(self._reg_queue.get())
+            except gevent.queue.Queue.Empty:
+                break
 
-            # This aggregator is non-functional
-            self.aggregator = self._get_api_href()
-            self.logger.writeInfo("Updated aggregator to {} (try {})".format(self.aggregator, i))
+        # Add items back to the queue
+        while True:
+            try:
+                self._reg_queue.put(new_queue.popleft())
+            except IndexError:
+                break
 
-        raise TooManyRetries(self._mdns_updater)
+    def register(self, res_type, key, **kwargs):
+        """Register 'resource' type data including the Node
+           NB: Node registration is managed by heartbeat thread so may take up to 5 seconds! """
+        self.register_into("resource", res_type, key, **kwargs)
+
+    def unregister(self, res_type, key):
+        """Unregister 'resource' type data including the Node"""
+        self.unregister_from("resource", res_type, key)
+
+    def register_into(self, namespace, res_type, key, **kwargs):
+        """General register method for 'resource' types"""
+        data = kwargs
+        send_obj = {"type": res_type, "data": data}
+        if 'id' not in send_obj["data"]:
+            self.logger.writeWarning("No 'id' present in data, using key='{}': {}".format(key, data))
+            send_obj["data"]["id"] = key
+
+        if namespace == "resource" and res_type == "node":
+            # Handle special Node type when Node is not registered, by immediately registering
+            if self._node_data["node"] is None:
+                # Will trigger registration in main thread
+                self._node_data["node"] = send_obj
+                return
+            # Update Node Data
+            self._node_data["node"] = send_obj
+        else:
+            self._add_mirror_keys(namespace, res_type)
+            self._node_data["entities"][namespace][res_type][key] = send_obj
+        self._queue_request("POST", namespace, res_type, key)
+
+    def unregister_from(self, namespace, res_type, key):
+        """General unregister method for 'resource' types"""
+        if namespace == "resource" and res_type == "node":
+            # Handle special Node type
+            self._unregister_node()
+            self._node_data["node"] = None
+            return
+        elif res_type in self._node_data["entities"][namespace]:
+            self._add_mirror_keys(namespace, res_type)
+            if key in self._node_data["entities"][namespace][res_type]:
+                del self._node_data["entities"][namespace][res_type][key]
+        self._queue_request("DELETE", namespace, res_type, key)
+
+    def _add_mirror_keys(self, namespace, res_type):
+        """Deal with missing keys in local mirror"""
+        if namespace not in self._node_data["entities"]:
+            self._node_data["entities"][namespace] = {}
+        if res_type not in self._node_data["entities"][namespace]:
+            self._node_data["entities"][namespace][res_type] = {}
+
+    def stop(self):
+        """Stop the Aggregator object running"""
+        self.logger.writeDebug("Stopping aggregator proxy")
+        self._running = False
+        self.main_thread.join()
+        self.queue_thread.join()
+
+    def status(self):
+        """Return the current status of node in the aggregator"""
+        return {"api_href": self.aggregator,
+                "api_version": self.aggregator_apiversion,
+                "registered": self._node_data["registered"]}
+
+    def _send(self, method, aggregator, api_ver, url, data=None):
+        """Handle sending request to the registration API, with error handelling
+        HTTP 200, 201, 204, 409 - Success, return response
+        Timeout, HTTP 5xx, Connection Error - Raise ServerSideError Exception
+        HTTP 4xx - Raise InvalidRequest Exception"""
+
+        url = "{}/{}/{}{}".format(AGGREGATOR_APINAMESPACE, AGGREGATOR_APINAME, api_ver, url)
+
+        try:
+            R = self._send_request(method, aggregator, url, data)
+            if R is None:
+                self.logger.writeWarning("No response from aggregator {}".format(aggregator))
+                raise ServerSideError
+
+            elif R.status_code in [200, 201, 204, 409]:
+                return R
+
+            elif (R.status_code // 100) == 4:
+                self.logger.writeWarning("{} response from aggregator: {} {}"
+                                         .format(R.status_code, method, urljoin(aggregator, url)))
+                raise InvalidRequest(R.status_code)
+
+            else:
+                self.logger.writeWarning("Unexpected status from aggregator {}: {}, {}"
+                                         .format(aggregator, R.status_code, R.content))
+                raise ServerSideError
+
+        except requests.exceptions.RequestException as e:
+            # Log a warning, then let another aggregator be chosen
+            self.logger.writeWarning("{} from aggregator {}".format(e, aggregator))
+            raise ServerSideError
+
+    def _send_request(self, method, aggregator, url, data=None):
+        """Low level method to send a HTTP request"""
+
+        self.logger.writeDebug("{} {}".format(method, urljoin(aggregator, url)))
+
+        # We give a long(ish) timeout below, as the async request may succeed after the timeout period
+        # has expired, causing the node to be registered twice (potentially at different aggregators).
+        # Whilst this isn't a problem in practice, it may cause excessive churn in websocket traffic
+        # to web clients - so, sacrifice a little timeliness for things working as designed the
+        # majority of the time...
+        if _config.get('prefer_ipv6') is False:
+            return requests.request(method, urljoin(aggregator, url), json=data, timeout=1.0)
+        else:
+            return requests.request(method, urljoin(aggregator, url), json=data, timeout=1.0, proxies={'http': ''})
 
 
 class MDNSUpdater(object):
-    def __init__(self, mdns_engine, mdns_type, mdns_name, mappings, port, logger, p2p_enable=False, p2p_cut_in_count=5,
+    def __init__(self, mdns_engine, mdns_type, mdns_name, mappings, port, logger, p2p_enable=False, p2p_cut_in_count=2,
                  txt_recs=None):
         self.mdns = mdns_engine
         self.mdns_type = mdns_type